Add preference to set domain.
I am hosting my own installation of Goread, but the android app
was using a hardcoded value for the domain. This is the first
rough draft of a preference to set a different URL.

One of the things I need to change is the login attempt right after
the first start of the app - this is still using the default value
for the domain "http://goread.io". I'm not quite sure yet how to
change that and maybe show the domain preference first...